+++ packages/cf-deploy-config-sub-generator/src/app/questions.ts
@@ -5,8 +5,8 @@ import {
     appRouterPromptNames,
     type CfAppRouterDeployConfigPromptOptions,
     type CfAppRouterDeployConfigQuestions,
-    type CfDeployConfigPromptOptions,
     type CfDeployConfigQuestions,
+    type CfDeployConfigPromptOptions,
     getAppRouterPrompts,
     getPrompts,
     promptNames
@@ -17,6 +17,7 @@ import { t } from '../utils';
 import type { ApiHubConfig } from '@sap-ux/cf-deploy-config-writer';
 import type { Answers, Question } from 'inquirer';
 import { withCondition } from '@sap-ux/inquirer-common';
+import type { Logger } from '@sap-ux/logger';
 
 /**
  * Fetches the Cloud Foundry deployment configuration questions.
@@ -63,12 +64,12 @@ export async function getCFQuestions({
             addBTPDestinationList: isBAS ? !isAbapDirectServiceBinding : false,
             additionalChoiceList: cfChoices
         },
-        [promptNames.addManagedAppRouter]: !mtaYamlExists && !isCap,
+        [promptNames.routerType]: !mtaYamlExists && !isCap,
         [promptNames.overwrite]: addOverwrite
     };
 
     DeploymentGenerator.logger?.debug(t('cfGen.debug.promptOptions', { options: JSON.stringify(options) }));
-    return getPrompts(options);
+    return getPrompts(options, DeploymentGenerator.logger as unknown as Logger);
 }
 
 /**