+++ packages/telemetry/src/base/client/azure-appinsight-client.ts
@@ -4,6 +4,7 @@ import { EventHeader } from '../types/event-header';
 import { SampleRate } from '../types/sample-rate';
 import { configAzureTelemetryClient } from '../utils/azure-client-config';
 import { TelemetrySettings } from '../config-state';
+import type { TelemetryMeasurements, TelemetryProperties } from '../types';
 
 /**
  *
@@ -48,8 +49,8 @@ class ApplicationInsightClient extends Client {
      */
     public reportBlocking(
         eventName: string,
-        properties: { [key: string]: string | boolean },
-        measurements: { [key: string]: number },
+        properties: TelemetryProperties,
+        measurements: TelemetryMeasurements,
         sampleRate: SampleRate | undefined,
         ignoreSettings: boolean = false
     ): Promise<void> {
@@ -74,8 +75,8 @@ class ApplicationInsightClient extends Client {
      */
     public async report(
         eventName: string,
-        properties: { [key: string]: string | boolean },
-        measurements: { [key: string]: number },
+        properties: TelemetryProperties,
+        measurements: TelemetryMeasurements,
         sampleRate: SampleRate | undefined,
         telemetryHelperProperties?: { [key: string]: string },
         ignoreSettings?: boolean
@@ -98,8 +99,8 @@ class ApplicationInsightClient extends Client {
      */
     private prepareClientAndEvent(
         eventName: string,
-        properties: { [key: string]: string | boolean },
-        measurements: { [key: string]: number },
+        properties: TelemetryProperties,
+        measurements: TelemetryMeasurements,
         sampleRate = SampleRate.NoSampling
     ): { client: appInsights.TelemetryClient; event: appInsights.Contracts.EventTelemetry } {
         const client = this.clients.get(sampleRate) as appInsights.TelemetryClient;