+++ packages/backend-proxy-middleware/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/backend-proxy-middleware
 
+## 0.8.40
+
+### Patch Changes
+
+-   Updated dependencies [28c6594]
+    -   @sap-ux/axios-extension@1.20.0
+
+## 0.8.39
+
+### Patch Changes
+
+-   Updated dependencies [d638daa]
+    -   @sap-ux/btp-utils@1.0.3
+    -   @sap-ux/axios-extension@1.19.3
+
 ## 0.8.38
 
 ### Patch Changes