+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/environment-check
 
+## 0.17.96
+
+### Patch Changes
+
+-   Updated dependencies [28c6594]
+    -   @sap-ux/axios-extension@1.20.0
+
+## 0.17.95
+
+### Patch Changes
+
+-   d638daa: Update @sap/bas-sdk dependency
+-   Updated dependencies [d638daa]
+    -   @sap-ux/btp-utils@1.0.3
+    -   @sap-ux/axios-extension@1.19.3
+
 ## 0.17.94
 
 ### Patch Changes